More stuff with the project
Now I can get the genes having a threshold which produces the greatest
information gain.

Unfortunately something seems to be wrong, since the maximal information
gain shown is 0. This probably means that the separation splitted the
examples into two sets: one of them is empty, the other is full of all
instances. This is totally weird.

I guess there's some bug into the code, I'd better check it out.